refactor(test-infra): keep TableBuilder file-local; fix stale canonical-loader rule text
